Update yearForm files and integrate MonthsGrid
- delete year-form-proposal folder
- changed some.css -> yearForm.css
- fixed months spent not updating when creating a new envelope
- finished patch year route to update each month with the corresponding values from MonthsGrid
