+++ main/client/src/components/yearForm.js
@@ -1,56 +1,99 @@
 import React, { useState, useEffect } from "react";
 import { useNavigate, useParams } from "react-router-dom";
+import MonthsGrid from './MonthsGrid';
+import '../yearForm.css';
 
 export default function YearForm (props) {
-  const [form, setForm] = useState({
-    year: "",
-    budget: "",
-  });
-  const params = useParams();
   const navigate = useNavigate();
+  const params = useParams();
+  const [year, setYear] = useState("");
+  const [remaining, setRemaining] = useState(0);
+  const [budget, setBudget] = useState("");
+  const [distribution, setDistribution] = useState(0);
+  const [months, setMonths] = useState(createMonthsObject(0, 0));
 
   useEffect(() => {
     if (params.year) {
       async function fetchData() {
-          
-        const response = await fetch(`http://localhost:5000/year/${params.year}`);
-            
+        const response = await fetch(
+          `http://localhost:5000/year/${params.year}`
+        );
+
         if (!response.ok) {
           const message = `An error has occurred: ${response.statusText}`;
           window.alert(message);
           return;
         }
-            
+
         const record = await response.json();
         if (!record) {
           window.alert(`Record with id ${params.year} not found`);
           navigate("/");
           return;
         }
-            
-        setForm(record);
+        const yearInfo = record[0];
+
+        setYear(yearInfo.year);
+        setBudget(yearInfo.budget);
+
+        const total = yearInfo.months.reduce((a, b) => a + b.budget, 0);
+        const leftover = yearInfo.budget - total;
+        setRemaining(leftover);
+        const currMonths = Object.fromEntries(yearInfo.months.map(month => [month.month, {budget: month.budget, max: month.budget + leftover}]));
+        setMonths(currMonths);
       }
-            
+
       fetchData();
-            
+
       return;
     }
   }, [params.year, navigate]);
-    
-  function updateForm(value) {
-    return setForm((prev) => {
-      return { ...prev, ...value };
-    });
+  
+  const updateYear = (e) => setYear(Number(e.target.value));
+
+  function getDistribution(n) {
+    const distributed = Math.floor(n / 12);
+    const remainder = n % 12;
+    return {distributed, remainder}
   }
+
+  function populateYears() {
+    const currentYear = new Date().getFullYear();
+    const maxYear = currentYear + 99;
+    let yearsArr = [];
+
+    for (let i = currentYear; i <= maxYear; i++) {
+      yearsArr.push((<option key={i} value={i}>{i}</option>));
+    }
+
+    return yearsArr;
+  }
+
+  function createMonthsObject(distributed, remainder) {
+    const monthsArr = ["January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"];
+    
+    const months = Object.fromEntries(monthsArr.map(month => [month, {budget: distributed, max: distributed}]));
+    months["January"].budget += remainder;
+    months["January"].max += remainder;
+    return months;
+  }
+
+  const updateBudget = (e) => {
+    const newBudget = e.target.value > 0 ? e.target.value : 0;
+    const {distributed, remainder} = getDistribution(newBudget);
+    setBudget(newBudget)
+    setRemaining(0)
+    setMonths(createMonthsObject(distributed, remainder))
+  };
 
   async function onSubmit(e) {
     e.preventDefault();
     const { isCreate } = props;
+    const newYear = { year, budget, months };
         
     if (isCreate) {
       // When a post request is sent to the create url, we'll add a new record to the database.
-      const newYear = { ...form };
-            
+      console.log(JSON.stringify(newYear));   
       await fetch("http://localhost:5000/year/add", {
         method: "POST",
         headers: {
@@ -62,21 +105,14 @@ export default function YearForm (props) {
           window.alert(error);
           return;
         });
-            
-      setForm({ year: "", budget: "" });
     } else {
-      const editedYear = {
-        year: form.year,
-        budget: form.budget,
-      };
-            
       // This will send a patch request to update the data in the database.
-      await fetch(`http://localhost:5000/year/${params.year}`, {
+      await fetch(`http://localhost:5000/year/${year}`, {
         method: "PATCH",
-        body: JSON.stringify(editedYear),
         headers: {
           'Content-Type': 'application/json'
         },
+        body: JSON.stringify(newYear),
       });
     }
         
@@ -84,49 +120,43 @@ export default function YearForm (props) {
   }
 
   return (
-    <form onSubmit={onSubmit}>
-      <div className="form-group">
-        <label htmlFor="year">Year</label>
-        <input
-          type="text"
-          className="form-control"
-          id="year"
-          minLength={4}
-          maxLength={4}
-          value={form.year}
-          onChange={(e) => updateForm({ year: e.target.value })}
-          required
-        />
+    <div className="year-creation">
+      <div className="container">
+        <form onSubmit={onSubmit}>
+          <div className="form-group">
+            <label htmlFor="year">Year</label>
+            <select
+              value={year}
+              className="form-control"
+              onChange={updateYear}
+              required
+            >
+              {populateYears()}
+            </select>
+          </div>
+          <div className="form-group">
+            <label htmlFor="budget">Budget</label>
+            <input
+              type="number"
+              className="form-control"
+              id="budget"
+              value={budget}
+              min={0}
+              onChange={updateBudget}
+              required
+            />
+          </div>
+          <div className="remaining">Remaining {remaining}</div>
+          <MonthsGrid distribution={distribution} budget={budget} months={months} setMonths={setMonths} remaining={remaining} setRemaining={setRemaining}/>
+          <div className="form-group">
+            <input
+              type="submit"
+              value={props.isCreate ? "Create Year" : "Edit Year"}
+              className="btn btn-primary"
+            />
+          </div>
+        </form>
       </div>
-      <div className="form-group">
-        <label htmlFor="budget">Budget</label>
-        <input
-          type="number"
-          className="form-control"
-          id="budget"
-          value={form.budget}
-          onChange={(e) => updateForm({ budget: e.target.value })}
-          required
-        />
-      </div>
-      <div className="form-group">
-        <label htmlFor="budget">Months</label>
-        <input
-          type="number"
-          className="form-control"
-          id="budget"
-          value={form.budget}
-          onChange={(e) => updateForm({ budget: e.target.value })}
-          required
-        />
-      </div>
-      <div className="form-group">
-        <input
-          type="submit"
-          value={props.isCreate ? "Create Year" : "Edit Year"}
-          className="btn btn-primary"
-        />
-      </div>
-    </form>
+    </div>
   );
 }
